backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
pydantic
pydantic-settings
langchain
langchain-community
langchain-core
langchain-openai
langchain-ollama
chromadb
networkx
pyvis
wikipedia
python-decouple
langchain-groq
python-dotenv
